// src/app/component/model/PurchaseElement.model.ts
export class PurchaseElement {
    constructor() {
    }
    purchaseId: string;
    postId: string;
    post: string;
    phoneNumber: string;
    address: string;
    sellerName: string;
    buyerName: string;
    unitPrice: number;
    purchaseNumber: number;
    dateOfOrder: Date;
    statusPurchase: string;
    imageURL: string;
    statusPurchaseName: string;
    transportCost: string;
}